requirements.txt
alembic==1.14.0
aniso8601==9.0.1
attrs==24.3.0
bcrypt==4.2.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.7
coverage==7.7.0
Deprecated==1.2.15
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.10.0
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
gunicorn==23.0.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
ldap3==2.9.1
limits==4.0.0
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.2
mdurl==0.1.2
mistune==3.1.0
ordered-set==4.1.0
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
Pygments==2.19.1
PyJWT==2.10.0
pytest==8.3.4
pytest-cov==6.0.0
pytest-dotenv==0.5.2
pytest-flask==1.3.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
six==1.17.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.1
